backend/requirements.txt
# FastAPI and dependencies
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.14.0

# Celery and Redis
celery==5.4.0
redis==5.2.1

# MinIO / S3
minio==7.2.11

# Media processing
yt-dlp==2024.12.23
ffmpeg-python==0.2.0

# Audio fingerprinting
pyacoustid==1.3.0

# Whisper (local)
faster-whisper==1.1.0

# OpenAI (for Whisper API)
openai==1.59.5

# Google Gemini
google-genai==1.41.0

# Embeddings
sentence-transformers==3.3.1

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.2
black==24.10.0
ruff==0.8.5
